* dump_multi_align: (29 commits)
  Updated the (short) doc about this pipeline
  --reg_conf is used by eHive, so let's use --registry to avoid any confusion
  The underscore is used in the MLSS name: the dot is a neater separator
  Also remove the existing files (leftover from a previous run)
  $tag should be an empty string by default
  Renamed "dump_mode" to "make_tar_archive"
  Gave a default value to "extra_args"
  Renamed "split_mode" to "split_by_chromosome" and rewrote some conditions
  The parameter is now consistently named "dump_mode"
  The "file" dump-mode needs to know how many blocks there are
  Updated EG config file
  Merged the code of InitInputID into the previous factory
  bugfix: filenames in emf2maf mode
  emf2maf must run before the compress job of the emf file
  Only use analysis_capacity throughtout this pipeline
  bugfix: the default parameters of the base-class must be included
  This part is in InitJobs
  The "species" parameter is not needed
  Only call "unlink" if there is a file
  Needs to include the options of the super-class
  ...

+++ modules/Bio/EnsEMBL/Compara/RunnableDB/DumpMultiAlign/DumpMultiAlign.pm
@@ -45,6 +45,17 @@ use strict;
 use warnings;
 use base ('Bio::EnsEMBL::Hive::RunnableDB::SystemCmd');
 
+sub param_defaults {
+    my $self = shift;
+    return {
+        %{ $self->SUPER::param_defaults() },
+        'cmd' => [ 'perl', '#dump_program#', '--species', '#species#', '--mlss_id', '#mlss_id#', '--masked_seq', '#masked_seq#', '--split_size', '#split_size#', '--output_format', '#format#', '--output_file', '#output_file_gen#' ],
+
+        'extra_args'    => [],
+    }
+}
+
+
 sub fetch_input {
     my $self = shift;
 
@@ -66,8 +77,8 @@ sub fetch_input {
 	push @$cmd, '--dbname', $self->param('compara_db');
     }
 
-    if ($self->param('reg_conf')) {
-	push @$cmd, '--reg_conf', $self->param('reg_conf');
+    if ($self->param('registry')) {
+	push @$cmd, '--reg_conf', $self->param('registry');
     }
 }
 
@@ -75,7 +86,7 @@ sub write_output {
     my $self = shift @_;
 
     #delete tmp file
-    unlink($self->param('tmp_file'));
+    unlink($self->param('tmp_file')) if $self->param('tmp_file');
 
     $self->SUPER::write_output();
 
@@ -90,7 +101,7 @@ sub write_output {
 sub _healthcheck {
     my ($self) = @_;
     
-    my $output_file = $self->param('output_file_pattern');
+    my $output_file = $self->param('output_file');
 
     my $cmd;
     if ($self->param('format') eq "emf") {
@@ -111,7 +122,7 @@ sub _healthcheck {
 	#visual confirmation all is well
 	my $sql = "INSERT INTO healthcheck (filename, expected,dumped) VALUES (?,?,?)";
 	my $sth = $self->db->dbc->prepare($sql);
-	$sth->execute($self->param('output_file_pattern'), $self->param('num_blocks'), $num_blocks);
+	$sth->execute($self->param('output_file'), $self->param('num_blocks'), $num_blocks);
 	$sth->finish();
     }
 }
